переносит создание слайдера в файл формы и добавляет ему валидацию
